Funds/requirements.txt
aiohttp==3.7.4.post0
ansible==2.9.19
appdirs==1.4.4
apturl==0.5.2
arrow==1.1.0
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
base58==2.1.0
bcrypt==3.1.7
bidict==0.21.2
binaryornot==0.4.4
bitarray==1.2.2
blinker==1.4
Brlapi==0.7.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
cleo==0.8.1
Click==7.0
clikit==0.6.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cookiecutter==1.7.2
crashtest==0.3.1
cryptography==2.8
cupshelpers==1.0
cytoolz==0.11.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gpg===1.13.1-unknown
hexbytes==0.2.2
html5lib==1.1
httplib2==0.14.0
idna==2.8
ipfshttpclient==0.8.0a2
isort==5.6.4
jeepney==0.6.0
Jinja2==2.11.3
jinja2-time==0.2.0
jsonschema==3.2.0
keyring==21.8.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
linecache2==1.0.0
llvmlite==0.35.0
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
meson==0.56.0
monotonic==1.5
msgpack==1.0.2
multiaddr==0.0.9
multidict==5.1.0
netaddr==0.8.0
netifaces==0.10.4
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
packaging==20.8
paramiko==2.6.0
parsimonious==0.8.1
pastel==0.2.1
pbr==5.5.1
pexpect==4.8.0
Pillow==7.0.0
pkginfo==1.7.0
poetry==1.1.4
poetry-core==1.0.0
poyo==0.5.0
protobuf==3.18.0
psutil==5.5.1
ptyprocess==0.7.0
py-solc-x==1.1.0
pycairo==1.16.2
pycrypto==2.6.1
pycryptodome==3.10.4
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylev==1.3.0
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.0
PySocks==1.6.8
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-engineio==4.0.1
python-magic==0.4.16
python-slugify==5.0.0
python-socketio==5.1.0
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rlp==2.0.1
rope==0.18.0
ropemode==0.5
ropevim==0.8.0
s3cmd==2.0.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.0
semantic-version==2.8.5
shellingham==1.3.2
simplejson==3.16.0
sip==4.19.21
six==1.14.0
stevedore==3.2.2
system-service==0.3
systemd-python==234
testresources==2.0.0
text-unidecode==1.3
toml==0.10.2
tomlkit==0.7.0
toolz==0.11.1
torbrowser-launcher==0.3.2
torch==1.7.1
traceback2==1.4.0
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
urllib3==1.25.8
usb-creator==0.3.7
varint==1.0.2
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
web3==5.23.1
webencodings==0.5.1
websocket-client==0.58.0
websockets==9.1
wrapt==1.12.1
xkit==0.0.0
yarl==1.6.3
youtube-dl==2020.3.24
